==> notebooks/4-evaluation(L4).ipynb <==
# Set Up

%%capture
import os
if "COLAB_" not in "".join(os.environ.keys()):
    !pip install unsloth
else:
    # Do this only in Colab notebooks! Otherwise use pip install unsloth
    !pip install --no-deps bitsandbytes accelerate xformers==0.0.29.post3 peft trl triton cut_cross_entropy unsloth_zoo
    !pip install sentencepiece protobuf "datasets>=3.4.1" huggingface_hub hf_transfer
    !pip install --no-deps unsloth

import sys
from google.colab import drive
import os

drive.mount('/content/drive')
drive_path = os.path.join('drive', 'MyDrive', 'Colab Notebooks', 'Dacon', 'SentenceOrder') # 다른 드라이브에서 사용할 경우, 이 부분만 적절히 수정

if drive_path not in sys.path:
  sys.path.append(drive_path)
from utility import *

SEED=42
CONFIG_PATH = drive_path + '/config.yaml'

set_all_seed(SEED)
config = load_config(CONFIG_PATH)

test = pd.read_csv(drive_path + '/dataset/test.csv').drop(columns = 'ID')

inference_message = [
    {"role": "system", "content": "You are an expert at understanding the logical flow of sentences. Your task is to arrange four given Korean sentences into a coherent and natural paragraph. Output the ordered sequence of sentence indices, separated by commas. /no_think"},
    {"role": "user", "content": "Provided Sentences:\n0. {sentence_0}\n1. {sentence_1}\n2. {sentence_2}\n3. {sentence_3}"}
]

# Evaluate

문장의 순서를 맞추는 작업에서는 창의성이 필요하지 않으므로 do_sample을 False로 두어 확률이 제일 높은 토큰을 생성

def evaluation(model, tokenizer, row, chat_template):
  # 하나의 테스트 데이터를 모델이 추론

  input = tokenizer(
      [ formatting_prompts(row, chat_template) ],
      return_tensors="pt"
  ).to("cuda")

  outputs = model.generate(
      input_ids=input.input_ids,
      attention_mask=input.attention_mask,
      eos_token_id=tokenizer.eos_token_id,
      max_new_tokens=32,
      use_cache=True,
      do_sample=False,
      temperature=None,
      top_p=None,
      top_k=None,
      min_p=None,
      num_beams=1,
  )
  response = tokenizer.batch_decode(outputs, skip_special_tokens=True)
  ans = list(map(int, response[0][-10:].split(','))) # 다른 텍스트를 전부 제외한, 순수 정답만 추출한 후 리스트로 변환

  return ans

checkpoint_list = [1200, 1400, 2400] # 추론에 사용할 모델의 체크 포인트 리스트

for check_point in checkpoint_list:
  # 체크 포인트 리스트에 있는 모든 체크 포인트에 대해, 모델을 불러온 후 추론하여 그 결과를 csv 파일로 저장

  sample_submission = pd.read_csv(drive_path + '/dataset/sample_submission.csv')

  model, tokenizer = FastLanguageModel.from_pretrained(
    model_name = drive_path + f'/models/qwen_14b_finetuning-r32-alpha32-unsloth/checkpoint-{check_point}',
    max_seq_length = 512,
    load_in_4bit = True,
  )

  model.eval()
  FastLanguageModel.for_inference(model)

  inference_chat_template = tokenizer.apply_chat_template(inference_message, tokenize=False, add_generation_prompt=True, enable_thinking=False)

  for i, row in tqdm(test.iterrows()):
    ans = evaluation(model, tokenizer, row, inference_chat_template)
    for j in range(4):
      sample_submission.iloc[i, 1 + j] = ans[j]
  sample_submission.to_csv(drive_path + f'/prediction/qwen_14b_finetuning-r32-alpha32-unsloth-{check_point}.csv')

  del model
  del tokenizer
  torch.cuda.empty_cache()
  gc.collect()

# Save the Best Model

# 제일 높은 성능을 가진 모델을 Hugging Face Model Hub에 저장
# 환경 변수에 HF_TOKEN을 설정해야 저장이 가능
# os.environ["HF_TOKEN"] = "hf_xxxxxxxxx"

check_point = 1200

model, tokenizer = FastLanguageModel.from_pretrained(
  model_name = drive_path + f'/models/qwen_14b_finetuning-r32-alpha32-unsloth/checkpoint-{check_point}',
  max_seq_length = 512,
  load_in_4bit = True,
)

new_model_name = "Qwen-3-14B-Sentence-Ordering"
model.push_to_hub(new_model_name)
tokenizer.push_to_hub(new_model_name)
