==> notebooks/6-ensemble(CPU).ipynb <==
import sys
from google.colab import drive
import os

drive.mount('/content/drive')
drive_path = os.path.join('drive', 'MyDrive', 'Colab Notebooks', 'Dacon', 'SentenceOrder')

if drive_path not in sys.path:
  sys.path.append(drive_path)
from utility import *

test = pd.read_csv(drive_path + '/dataset/test.csv')
answers_list = [dict() for _ in range(len(test))]

for cp in tqdm([1200, 1400, 2400]):
  qwen_fine_tuning = pd.read_csv(drive_path + f'/prediction/qwen_14b_finetuning-r32-alpha32-unsloth-{cp}.csv')
  for i, row in qwen_fine_tuning.iterrows():
    string = f"{row['answer_0']}, {row['answer_1']}, {row['answer_2']}, {row['answer_3']}"
    if string in answers_list[i]:
      answers_list[i][string] += 1
    else:
      answers_list[i][string] = 1

sample_submission = pd.read_csv(drive_path + '/dataset/sample_submission.csv')
for i, ans in enumerate(answers_list):
  ans = max(answers_list[i], key=answers_list[i].get).split(', ')
  for j in range(4):
    sample_submission.iloc[i, 1 + j] = np.int64(ans[j])

sample_submission.to_csv(drive_path + f'/prediction/qwen_14b_final_ensemble.csv')
